fix(create-plugin): make sure upgrade and migrate export correct filenames
